remove livenessProbe from manga statefulsets
The livenessProbes would kill the pod, but then try to stat the volume at
a kubelet location attached to the EXISTING pod id.  That doesn't
work.  The file handle will still be stale.

What does work is deleting the pod and having the statefulset controller
bring up a new one.  That ends up generating a fresh pod id, and hence,
a new volume attachment for the NFS volume.